Move Categorical outcome selector to Sidebar
