+++ app/src/Sidebar/CategoricalSidebar/Buy/OutcomeSelection.js
