+++ app/src/Sidebar/CategoricalSidebar/Buy/OutcomeSelection.scss
