+++ app/src/Sidebar/CategoricalSidebar/Buy/buy.scss
@@ -193,7 +193,7 @@
   width: 100%;
   cursor: pointer;
   color: $white;
-  background: $green;
+  background: $blue;
   border-radius: .2rem;
   height: 5.2rem;
   padding: 0 1rem;
