+++ app/src/Sidebar/CategoricalSidebar/sidebar.scss
@@ -4,7 +4,6 @@
   box-shadow: 0 0 26px rgba(0, 25, 60, 0.12);
   height: 100%;
   width: 100%;
-  padding: 0 0 10rem;
   display: table;
 
   @media #{$tablet-small-max} {
